rename load-course-data back to load-data
